+++ AdminSite/src/adminsite/adminsite/site/employees/hrm/timesheetemployee.aspx.cs
@@ -26,7 +26,7 @@ namespace adminsite.site.employees.hrm
                         string timesheetString = (string)Session["CONSULTED_TIMESHEET_HRM"];
                         if (timesheetString != null)
                         {
-                            Timesheet timesheet = new Timesheet(Int32.Parse(timesheetString));
+                            Timesheet timesheet = new Timesheet(Int64.Parse(timesheetString));
                             GetAllWorkloadsByTimesheetCommand cmd = new GetAllWorkloadsByTimesheetCommand(timesheet);
                             cmd.Execute();
                             timesheet = cmd.GetResults();
@@ -204,7 +204,7 @@ namespace adminsite.site.employees.hrm
             try
             {
                 string timesheetString = (string)HttpContext.Current.Session["CONSULTED_TIMESHEET_HRM"];
-                Timesheet timesheet = new Timesheet(Int32.Parse(timesheetString));
+                Timesheet timesheet = new Timesheet(Int64.Parse(timesheetString));
                 GetAllWorkloadsByTimesheetCommand cmd = new GetAllWorkloadsByTimesheetCommand(timesheet);
                 cmd.Execute();
                 timesheet = cmd.GetResults();
@@ -272,7 +272,7 @@ namespace adminsite.site.employees.hrm
                 if (active)
                 {
                     string timesheetString = (string)Session["CONSULTED_TIMESHEET_HRM"];
-                    Timesheet timesheet = new Timesheet(Int32.Parse(timesheetString));
+                    Timesheet timesheet = new Timesheet(Int64.Parse(timesheetString));
                     timesheet.status = status;
                     UpdateTimesheetStatusCommand cmd = new UpdateTimesheetStatusCommand(timesheet);
                     cmd.Execute();
